requirements.txt
# Core
anthropic==0.79.0
python-dotenv==1.0.0
pydantic==2.12.5

# Google Sheets
gspread==6.2.1
google-auth==2.48.0
google-auth-oauthlib==1.2.4
google-auth-httplib2==0.3.0

# Claude API / HTTP
httpx==0.28.1
requests==2.32.5

# SNS
tweepy==4.16.0

# Templates
Jinja2==3.1.6

# Data / Config
PyYAML==6.0.3

# Web scraping & form automation
beautifulsoup4==4.13.3
playwright==1.50.0

# Google Analytics
google-analytics-data>=0.18.0

# Google Ads
google-ads>=25.0.0

# Google Cloud Storage
google-cloud-storage>=2.18.0

# Gemini (fallback)
google-generativeai>=0.8.0
